reducer put in reducers/application.js with test passing, all other tests passing
