+++ src/reducers/application.test.js
@@ -1,0 +1,9 @@
+import reducer from "reducers/application";
+
+describe("Application Reducer", () => {
+  it("thows an error with an unsupported type", () => {
+    expect(() => reducer({}, { type: null })).toThrowError(
+      /tried to reduce with unsupported action type/i
+    );
+  });
+});
